code-repetition reduced
